-  enhance home feature and some bugs in ui - implement bottom nav app bar using statefullShellBranch - handle tests to be full passed

+++ lib/features/home/domain/entities/portfolio_snapshot_entity.dart
@@ -1,6 +1,8 @@
+import 'package:decimal/decimal.dart';
+
 class PortfolioSnapshotEntity {
   const PortfolioSnapshotEntity({required this.date, required this.realTotal});
 
   final DateTime date;
-  final double realTotal;
+  final Decimal realTotal;
 }
